+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Optimized import times for models and maps with many materials and images.
 * Adjusted images and UVs to have the expected vertical orientation.
 
+### Fixed
+* Fixed an issue where some animation files failed to load.
+
 ## 0.3.1 - 2024-04-19
 ### Fixed
 * Fixed an issue where map files failed to load.