+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 0.4.0
+## 0.4.0 - 2024-04-30
 ### Added
 * Added an option to save unpacked PNG textures when importing to reduce memory usage and import times.